scribe/skills/scripts/pyproject.toml
[project]
name = "scribe-skill"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0.0",
]
